+++ Dolphin/Source/Core/ObjectBuilder/BuilderChain/BuilderChain_member.h
@@ -1,6 +1,6 @@
 #pragma once
 #include "Source/Core/Object/Object_forward.h"
-#include "Source/StandardComponent/Nest.h"
+#include "StandardComponent/Nest.h"
 #include "BuilderChain_forward.h"
 
 namespace Dolphin
